build(deps): bump coti-ethers/sdk; pin OZ 5.4.0
Signed setPublic256 overloads made bare 0 ambiguous.
Pin OZ so ^ does not float past 0.8.20. Fall back to
SIGNING_KEYS when PRIVATE_KEY unset for Hardhat accounts.

+++ contracts/pod/token/perc20/cotiside/PodErc20CotiMother.sol
@@ -330,7 +330,7 @@ contract PodErc20CotiMother is IPodErc20CotiSide, InboxUser, Ownable {
     // --- Private: garbled balance helpers ---
 
     function _ciphertextPlainZero() private returns (ctUint256 memory) {
-        return MpcCore.offBoard(MpcCore.setPublic256(0));
+        return MpcCore.offBoard(MpcCore.setPublic256(uint256(0)));
     }
 
     function _isEmptyCtUint256(ctUint256 memory ct) private pure returns (bool) {
